requirements.txt
panda3d
panda3d-simplepbr==0.12.0
